+++ modules/control_system/functions/@ss/properties.m
@@ -1,0 +1,31 @@
+%=============================================================================
+% Copyright (c) 2023-present Allan CORNET (Nelson)
+%=============================================================================
+% This file is part of the Nelson.
+%=============================================================================
+% LICENCE_BLOCK_BEGIN
+% SPDX-License-Identifier: LGPL-3.0-or-later
+% LICENCE_BLOCK_END
+%=============================================================================
+function varargout = properties(obj)
+  props = {'A'; 'B'; 'C'; 'D'; 'E'; 'Ts'; 'TimeUnit'; 'UserData'};
+  if (nargout == 0)
+    currentFormat = format();
+    if strcmp(currentFormat.LineSpacing, 'loose')
+      disp(' ');
+    end
+    disp('Properties for class tf:');
+    if strcmp(currentFormat.LineSpacing, 'loose')
+      disp(' ');
+    end
+    for p = props'
+      disp(['    ', p{1}]);
+    end
+    if strcmp(currentFormat.LineSpacing, 'loose')
+      disp(' ');
+    end
+  else
+    varargout{1} = props;
+  end
+end
+%=============================================================================